Fix patched image generation
dd did a partial read of 64KB (pipe buffer size), but doesn't warn with count=1...
Partial reads cannot be avoided with pipes without the non-POSIX iflag=fullblock.
With conv=sync, nul bytes are added on partial reads.
Without conv=sync, a smaller output would be produced.

Now we use dd to seek at the desired size, and write from /dev/null (ie 0 byte).
This is not fully standard, but it works (and at worst it shouldn't corrupt the file).